Finish reading the entire file in iso9660
size here is the amount we have left to write. To determine if we
are done, we need to compare the new offset to the total size of
the file.

Without this fix we would always bail at around halfway
(+/- the buffer size) through any file larger than the buffer size.