Check the filepath.WalkFunc error in iso9660.Finalize
If there is an error getting the file information at any point in the
tree info will be nil and err will be set. The WalkFunc must decide how
to handle the passed error.

In this case, as we can't finalize the iso, we choose to error out.

Before this fix, a failure to access some file in the tree would cause a
panic.