Make FAT32 label read consistent with Windows
While there are 2 possibles locations for the FAT32 label, in practice,
Windows only considers the special file in the root directory and
completely ignores the bootsector. For consistency purposes, Linux'
``fatlabel`` decided to follow this behavior as well. To avoid any
quirks, this commit moves the label read to the special file.

This should not pose compatibility issue with images previously
generated with this library since it was already setting the label in
both possible locations.