Add support for setting a FAT32 label (#154)
* Extract fat structure writes out of the Create function

The addition of ``SetLabel()`` will require to re-use some of the write
path currently only in the ``Create()`` function. To avoid code
duplication and for consistency, this function extract the various
internal structures write paths to dedicated functions.

* Add function to set filesystem label

This commit adds ```SetLabel()` to the filesystem interface. For
read-only filesystems, it returns an error.

For FAT32, ``SetLabel()`` writes the label in the boot sector, the
backup bootsector and the special file in the root directory.

* Make FAT32 label read consistent with Windows

While there are 2 possibles locations for the FAT32 label, in practice,
Windows only considers the special file in the root directory and
completely ignores the bootsector. For consistency purposes, Linux'
``fatlabel`` decided to follow this behavior as well. To avoid any
quirks, this commit moves the label read to the special file.

This should not pose compatibility issue with images previously
generated with this library since it was already setting the label in
both possible locations.

* Add FAT32 SetLabel() tests