squashfs: fix directory listings (#195)
This fixes directory listings sometimes giving the error:

    could not parse directory header: header was 3 bytes, less than minimum 12

This was caused by a check for "at least one directory header" failing
on empty directories.

On careful examination of the spec, it says that if a directory header
is present it must have at least one directory entry, but it is silent
on whether a directory header has to exist or not. A non-existent
header seems like the only way to sensibly implement empty
directories.

This adds a check to implement the "at least one directory entry per
directory header" check and modifies the bounds check to catch a
header with no entries.

This also introduces a test checking that we can recurse the test
squashfs file correctly - this fails without this patch.