squashfs: fix received X bytes instead of expected minimal 40 on listing
When unpacking large squashfs archives you occasionally get this
error:

    received 38 bytes instead of expected minimal 40

This is due to the inode changing type (ie it is actual an extended
inode) when read from the directory but the size not being updated.
Most of the time there is enough data for the larger inode but not
always.