MBR: add partition table UUID and partition UUID (#198)
This allows to retrieve the partition table UUID and partition UUID for
MBR formatted disks. The partition table UUID (aka "disk identifier" or
"disk signature") can be used to identify a disk when using the MBR
format scheme. The partition UUID is then just the partition table UUID
with the partitions index number appended.