squashfs: fix fragment block location
The granularity of fragment data writing is block, so we don't need to
return an unaligned fragment amount but 'sized chunks'

everytime we write a fragmentdata block, the variable location should be
added with a size of 'blocksize', which means that we have written a new
fragment block.

The variable `location` will be reused to tell new accumulated
fragmentblocks position.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>