Squashfs: fix error format when file larger than block size (#287)
* squashfs: fix blocks start field in FileInode

Due to SquashFS, the first field of both File inode and extended file
inode is called 'blocks start', that means

```
The offset from the start of the archive to the first data block.
```

Link https://dr-emann.github.io/squashfs/squashfs.html#_file_inodes

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>

* squashfs: fix fragment block location

The granularity of fragment data writing is block, so we don't need to
return an unaligned fragment amount but 'sized chunks'

everytime we write a fragmentdata block, the variable location should be
added with a size of 'blocksize', which means that we have written a new
fragment block.

The variable `location` will be reused to tell new accumulated
fragmentblocks position.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>

* squashfs: clear cumulative variable for fragment data

The variable fragmentData is used to record if the current fragment data
is enough to occupy a block. Once a block is written, it should be
cleared for a new round of loop.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>

* squashfs: set file inode's frag index when no fragment data

Due to spec, frag index of file inode

https://dr-emann.github.io/squashfs/squashfs.html#_file_inodes

```
An index into the Fragment Table which describes the fragment block
that the tail end of this file is stored in. If not used, this is set
to 0xFFFFFFFF.
```

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>

* lint: fix go fmt lint error

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>

---------

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>