Fix vds allocation
with `vds := make([]volumeDescriptor, 2)` we have
slice with two nil values, to which we later
append `vds = append(vds, vd)`, so PVD ends up to be
third element with SVDs appended after it.

Specs said there could be up to 127 SVDs, so,
let's allocate slice with 0 elements and capacity 128: PVD + 127 SVDs.